requirements.txt
Django
djangorestframework
psycopg2-binary
django-two-factor-auth
django-otp